fix: persist profiles in localStorage
