+++ src/admin.js
@@ -1,53 +1,4 @@
-class ProfileStorage {
-    static STORAGE_KEY = 'commission_profiles';
-
-    static saveProfiles(profiles) {
-        const data = {
-            version: '1.0',
-            timestamp: Date.now(),
-            profiles: profiles
-        };
-        localStorage.setItem(this.STORAGE_KEY, JSON.stringify(data));
-    }
-
-    static loadProfiles() {
-        try {
-            const stored = localStorage.getItem(this.STORAGE_KEY);
-            if (!stored) return null;
-            const data = JSON.parse(stored);
-            return data.profiles || null;
-        } catch (e) {
-            console.error('Error cargando perfiles:', e);
-            return null;
-        }
-    }
-
-    static exportProfiles() {
-        const profiles = this.loadProfiles() || DEFAULT_PROFILES;
-        const blob = new Blob([JSON.stringify(profiles, null, 2)], { type: 'application/json' });
-        const url = URL.createObjectURL(blob);
-        const a = document.createElement('a');
-        a.href = url;
-        a.download = `perfiles_comisiones_${new Date().toISOString().split('T')[0]}.json`;
-        a.click();
-    }
-
-    static importProfiles(file) {
-        return new Promise((resolve, reject) => {
-            const reader = new FileReader();
-            reader.onload = (e) => {
-                try {
-                    const profiles = JSON.parse(e.target.result);
-                    this.saveProfiles(profiles);
-                    resolve(profiles);
-                } catch (err) {
-                    reject(err);
-                }
-            };
-            reader.readAsText(file);
-        });
-    }
-}
+import { ProfileStorage } from "./storage.js";
 
 const DEFAULT_BASE_CONFIG = {
     niveles: ['Capilla', 'Junior', 'Senior A', 'Senior B', 'Máster', 'Genio'],
