+++ src/comisiones.css
@@ -1,3 +1,4 @@
+:root{--primary:#006D77;--success:#4CAF50;--warning:#FFA726;--danger:#F44336;}
 body {
     margin: 0;
     font-family: Arial, sans-serif;
@@ -48,12 +49,15 @@ input:not(:placeholder-shown) {
     font-weight: bold;
 }
 .progress {
+    border-radius: 8px;
+    overflow: hidden;
+    box-shadow: 0 2px 4px rgba(0,0,0,0.1);
     display: flex;
     height: 12px;
     margin-bottom: 5px;
 }
 .progress div {
-    flex: 1;
+    transition: background 0.3s, transform 0.3s;
     background: #ddd;
     margin-right: 2px;
     cursor: pointer;
@@ -62,8 +66,10 @@ input:not(:placeholder-shown) {
     margin-right: 0;
 }
 .progress div.reached {
-    background: #4caf50;
+    background: var(--success);
 }
 .progress div.current {
-    background: #003f7f;
+    background: var(--primary);
 }
+.progress div:hover{transform:scale(1.05);}
+
