+++ src/index.html
@@ -74,6 +74,6 @@
         <section id="detalle"></section>
     </main>
 </div>
-<script src="app.js"></script>
+<script type="module" src="app.js"></script>
 </body>
 </html>
